requirements.txt
python-dotenv==1.0.0
streamlit==1.31.1
openai==1.12.0
pandas==2.2.0
